+++ src/core/mod.rs
@@ -14,6 +14,7 @@ pub mod agent_task;
 pub mod agent_task_aggregate;
 pub mod agent_task_controller_service;
 pub mod agent_task_cook_loop;
+pub mod agent_task_dispatch_service;
 pub mod agent_task_fanout;
 pub mod agent_task_finalization;
 pub mod agent_task_gate;
@@ -63,6 +64,7 @@ pub mod hygiene;
 pub(crate) mod io;
 pub mod issues;
 pub mod keychain;
+pub mod lab_routing;
 pub mod observation;
 pub mod output;
 pub(crate) mod ownership;
